No longer need to resize rhcos image for MOC
The latest rhcos image are based on RHEL8 and have a different
partition layout than the previous ones based on RHEL7. They now
include an xfs partition that make they difficult to shrink as we did
before. Luckily MOC provided us with flavors with 25GB drives to
accomodate with the image size.