Decouple add of LB and apps FIPs
It is preferable in CI to only create the apps FIP once the deployment
finishes, as our FIP quota is rather small. This means we can't update
the CI DNS records for LB and APPS with a single curl call.

This change makes it possible to add records for LB FIP and apps FIP in
two separated calls, while ensuring the iptables rule is added only
once.