Make builders/build_capo.sh working out of the box
This will help to build openshift/capo without complication.

You just need a valid auth file and that's it.
If you want to specify a directory for CAPO, use $CAPO_DIR and it won't
clone it.

Also it adds some validation at the end to make sure that the release
image is usable for an actual deployment.

Signed-off-by: Emilien Macchi <emilien@redhat.com>