List and show stale ports
Previously, the `stale` script relied on the ability of Neutron to
return the update_date right from the `list` command.

It looks like this capability is not always present.

With this change, to list stale ports this script does not leverage
shortcuts any more, and instead lists and shows the date in separate
steps.