clean-ci-resources: mark undeletable volumes as 'error'
If volumes can't be removed because of their current state, they first need to be set as 'error' and then later they can be removed.

This patch will add a check if the resource is a volume and it not in the right state,
we mark it as error, so later it can be removed.