Add a script to report borderline quota usage
This script can be used in CI to report if some of our resources are
closed to reach the maximum quota limit.

By default, it reports if a resource has less than 15% free.

It relies on "openstack limits show --absolute" and it was tested
against multiple clouds: PSI, VEXXHOST managed, mecha.