Containerfile: hack to have more reliable builds
Fedora repos are not reliable, and very often we have issues when
building the images, where the repo is not available, under heavy use
and therefore can't serve the packages that we need.

This is problematic as it's causing some of our CI jobs to fail, and
creating a lot of noise.

To address this issue, we propose to implement an hack to safely run
DNF with retries (up to 10 retries when a failure is hit, with 5s of
break between attempts).

This patch also adds python3-openstackclient & python3-octaviaclient
& jq which are needed by our CI job running shiftstack-ci specific
jobs.