server: use --wait when creating LB resources
To avoid race condition when we create a LB resource that depends on
a parent resource, we can use `--wait` argument so the resource is
created for sure when the commands finishes, before we go to the next
command that creates a child resource.

That will help us to avoid things like:

```
Load Balancer 4b87150b-e5a9-4cb7-b34f-e3196a9f5bdc is immutable and cannot be updated. (HTTP 409) (Request-ID: req-f04bcbab-b9eb-40ec-b7ea-eb716ec91ad9)
```