clean-ci-resources: tolerate failures in first attempt
If we tolerate cluster destroy failures in the first attempt, that uses
openshif-install, we'll give a chance to the other attempts to clean up
staled resources.
The second attempt is running with --force, which uses the openstackcli.
And finally, we have a loop that clean some specific resources.

This has been tried against VEXXHOST managed when we had some recurring
networking resources in staled, and they were cleaned.

This patches declares an initial CLEANUP_RC set to 0, then increases the
RC every time that a command is run. At the end we check the value for
$CLEANUP_RC, so if a failure was found somewhere, we'll report it in the
logs but resources will be gone.

Note: the cleanup CI job will report a failure, but if the resources
were removed in one of the attempts, then the cleanup script won't fail
anymore since the resources are gone.
Comparing to what we have today, where we have the job running every
2 hours and complaining about the first attempt which fails to remove
  the resources...