destroy_cluster: run cluster destroy with a timeout
This script is run in our CI and the job currently timeouts if
`openshift-install destroy` never manage to finish (it's the case when a
resource is not destroyed). So let's run the destroy command during 15
minutes and if it didn't finish on time, report it in the logs but move
on.
Later, we have code that will force the deletion of these ressources.