Make metrics more prometheus idiomatic
As the metric is an openstack quota we put openstack_quota in the name.
The name becomes openstack_quota_<service>_<resource>.

As quota is now in the metric name we also change the 'quota' label to
'type'.

We include the cloud and projectid in the metric.

Include parseable prometheus metadata in the output.