Make the liveliness periodic more robust
The previous SSH test was prone to sporadic failures with a failure rate
above 10% for the amphora driver in vh-mecha-central. The HTTP test has
been reliable so far in testing.

The liveliness periodic invokes server.sh with `-u centos`, which both
supplied the username to use when logging in via SSH and enabled the
test. We no longer require a login username for the test so we add a new
option `-l` which enables the test without taking an option. We retain
the `-u` option for compatibility with the periodic. It can be removed
after we have updated the periodic.