Add team_ssh_keys.sh
This script will generate a temporary file with all the ssh keys of the
approvers of a given group in the OWNERS_ALIASES file of a given project.
The script will return the path of the temporary file.
The script will return 1 if the script fails to curl the OWNER_ALIASES file
or if the yq command is not found.