Update bump-periodics.sh
The path of the config directory changed in the release repository.

Copy TP jobs as well.

Drop code that would remove out of support releases. This is not
reliable with newer long term releases so we do the cleanup manually.

Also, point all maintenance jobs to nightly (TP and upgrade, and not
just the main jobs).

And finally, do not update the interval as these were not consistent and
are likely better left off for now.